backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
openai==1.2.4
apprise==1.4.5
python-dotenv==1.0.0
apscheduler==3.10.4